Whisper/requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart
requests
